refactor(states): move around and rename things
