+++ core/src/states/task.rs
@@ -1,163 +1,128 @@
-use crate::{
-    bitcoin_syncer::BitcoinSyncerEvent, builder::transaction::OperatorData, database::Database,
-};
+use crate::{bitcoin_syncer::BitcoinSyncerEvent, database::Database};
 use pgmq::{Message, PGMQueueExt};
 use std::time::Duration;
 use tokio::{sync::oneshot, task::JoinHandle};
 
-use crate::{
-    builder::transaction::DepositData, config::protocol::ProtocolParamset,
-    database::DatabaseTransaction, errors::BridgeError, rpc::clementine::KickoffId,
-    states::SystemEvent,
-};
+use crate::{config::protocol::ProtocolParamset, errors::BridgeError, states::SystemEvent};
 
 use super::{context::Owner, StateManager};
 
-/// Starts a new thread to periodically fetch new blocks from bitcoin_syncer
-pub async fn fetch_new_blocks<T: Owner + std::fmt::Debug + 'static>(
-    last_processed_block_height: u32,
-    db: Database,
-    poll_delay: Duration,
-    paramset: &'static ProtocolParamset,
-) -> JoinHandle<Result<(), eyre::Report>> {
-    tokio::spawn(async move {
-        let queue_name = StateManager::<T>::queue_name();
-        tracing::info!(
-            "Starting state manager block syncing with owner type {} starting from height {}",
-            T::OWNER_TYPE,
-            last_processed_block_height
-        );
-        // variable to store locally last sent height
-        let mut last_sent_height = last_processed_block_height;
-        let mut num_consecutive_errors = 0;
-        let queue = PGMQueueExt::new_with_pool(db.get_pool()).await;
-        loop {
-            let result: Result<bool, eyre::Report> = async {
-                let mut dbtx = db.begin_transaction().await?;
-                let has_updated_states = async {
-                    let event = db.get_event_and_update(&mut dbtx, &queue_name).await?;
-                    Ok::<bool, eyre::Report>(match event {
-                        Some(event) => match event {
+impl<T: Owner + std::fmt::Debug + 'static> StateManager<T> {
+    /// Starts a new thread to periodically fetch new blocks from bitcoin_syncer
+    pub async fn block_fetcher_task(
+        last_processed_block_height: u32,
+        db: Database,
+        poll_delay: Duration,
+        paramset: &'static ProtocolParamset,
+    ) -> JoinHandle<Result<(), eyre::Report>> {
+        tokio::spawn(async move {
+            let queue_name = Self::queue_name();
+            tracing::info!(
+                "Starting state manager block syncing with owner type {} starting from height {}",
+                T::OWNER_TYPE,
+                last_processed_block_height
+            );
+
+            // variable to store locally last sent height
+            let mut last_sent_height = last_processed_block_height;
+            let mut num_consecutive_errors = 0;
+            let queue = PGMQueueExt::new_with_pool(db.get_pool()).await;
+            loop {
+                let result: Result<bool, eyre::Report> = async {
+                    let mut dbtx = db.begin_transaction().await?;
+                    let has_updated_states = async {
+                        let Some(event) = db.get_event_and_update(&mut dbtx, &queue_name).await?
+                        else {
+                            return Ok(false);
+                        };
+
+                        let new_tip = match event {
                             BitcoinSyncerEvent::NewBlock(block_id) => {
                                 let current_tip_height = db
                                     .get_block_info_from_id(Some(&mut dbtx), block_id)
                                     .await?
                                     .ok_or(eyre::eyre!("Block not found"))?
                                     .1;
+
                                 let mut new_tip = false;
+
                                 // update states to catch up to finalized chain
                                 while last_sent_height
                                     < current_tip_height - paramset.finality_depth + 1
                                 {
+                                    new_tip = true;
+
                                     let next_height = last_sent_height + 1;
-                                    let block =
-                                        db.get_full_block(Some(&mut dbtx), next_height).await?;
-                                    if let Some(block) = block {
-                                        new_tip = true;
-                                        let event = SystemEvent::NewBlock {
-                                            block,
-                                            height: next_height,
-                                        };
-                                        queue
-                                            .send_with_cxn(&queue_name, &event, &mut *dbtx)
-                                            .await
-                                            .map_err(|e| {
-                                                BridgeError::Error(format!(
-                                                    "Error sending event: {:?}",
-                                                    e
-                                                ))
-                                            })?;
-                                    } else {
-                                        return Err(eyre::eyre!(format!(
+
+                                    let block = db
+                                        .get_full_block(Some(&mut dbtx), next_height)
+                                        .await?
+                                        .ok_or(eyre::eyre!(format!(
                                             "Block at height {} not found",
                                             next_height
-                                        )));
-                                    }
+                                        )))?;
+
+                                    let event = SystemEvent::NewBlock {
+                                        block,
+                                        height: next_height,
+                                    };
+
+                                    queue
+                                        .send_with_cxn(&queue_name, &event, &mut *dbtx)
+                                        .await
+                                        .map_err(|e| {
+                                            BridgeError::Error(format!(
+                                                "Error sending event: {:?}",
+                                                e
+                                            ))
+                                        })?;
                                     last_sent_height += 1;
                                 }
+
                                 new_tip
                             }
                             BitcoinSyncerEvent::ReorgedBlock(_) => false,
-                        },
-                        None => false,
-                    })
-                }
-                .await?;
-                dbtx.commit().await?;
-                if has_updated_states {
-                    // Don't wait in new events
-                    return Ok(true);
-                }
+                        };
 
-                Ok(false)
-            }
-            .await;
-
-            match result {
-                Ok(true) => {
-                    num_consecutive_errors = 0;
-                }
-                Ok(false) => {
-                    num_consecutive_errors = 0;
-                    tokio::time::sleep(poll_delay).await;
-                }
-                Err(e) => {
-                    tracing::error!("State manager block fetch error: {:?}", e);
-                    num_consecutive_errors += 1;
-                    if num_consecutive_errors > 50 {
-                        return Err(eyre::eyre!(
-                            "Too many consecutive state machine block fetching errors"
-                        ));
+                        Ok::<bool, eyre::Report>(new_tip)
                     }
-                    tokio::time::sleep(poll_delay).await;
+                    .await?;
+
+                    dbtx.commit().await?;
+
+                    if has_updated_states {
+                        // Don't wait in new events
+                        return Ok(true);
+                    }
+
+                    Ok(false)
+                }
+                .await;
+
+                match result {
+                    Ok(true) => {
+                        num_consecutive_errors = 0;
+                    }
+                    Ok(false) => {
+                        num_consecutive_errors = 0;
+                        tokio::time::sleep(poll_delay).await;
+                    }
+                    Err(e) => {
+                        tracing::error!("State manager block fetch error: {:?}", e);
+                        num_consecutive_errors += 1;
+                        if num_consecutive_errors > 50 {
+                            return Err(eyre::eyre!(
+                                "Too many consecutive state machine block fetching errors"
+                            ));
+                        }
+                        tokio::time::sleep(poll_delay).await;
+                    }
                 }
             }
-        }
-    })
-}
+        })
+    }
 
-pub async fn add_new_round_machine<T: Owner + std::fmt::Debug + 'static>(
-    db: Database,
-    tx: DatabaseTransaction<'_, '_>,
-    operator_data: OperatorData,
-    operator_idx: u32,
-) -> Result<(), eyre::Report> {
-    let queue_name = StateManager::<T>::queue_name();
-    let queue = PGMQueueExt::new_with_pool(db.get_pool()).await;
-    let message = SystemEvent::NewOperator {
-        operator_data,
-        operator_idx,
-    };
-    queue
-        .send_with_cxn(&queue_name, &message, &mut *(*tx))
-        .await
-        .map_err(|e| BridgeError::Error(format!("Error sending NewOperator event: {:?}", e)))?;
-    Ok(())
-}
-
-pub async fn add_new_kickoff_machine<T: Owner + std::fmt::Debug + 'static>(
-    db: Database,
-    tx: DatabaseTransaction<'_, '_>,
-    kickoff_id: KickoffId,
-    kickoff_height: u32,
-    deposit_data: DepositData,
-) -> Result<(), eyre::Report> {
-    let queue_name = StateManager::<T>::queue_name();
-    let queue = PGMQueueExt::new_with_pool(db.get_pool()).await;
-    let message = SystemEvent::NewKickoff {
-        kickoff_id,
-        kickoff_height,
-        deposit_data,
-    };
-    queue
-        .send_with_cxn(&queue_name, &message, &mut *(*tx))
-        .await
-        .map_err(|e| BridgeError::Error(format!("Error sending NewKickoff event: {:?}", e)))?;
-    Ok(())
-}
-
-impl<T: Owner + std::fmt::Debug + 'static> StateManager<T> {
-    pub async fn into_polling_task(
+    pub async fn into_msg_consumer_task(
         mut self,
         poll_delay: Duration,
     ) -> (JoinHandle<Result<(), eyre::Report>>, oneshot::Sender<()>)
@@ -292,7 +257,7 @@ mod tests {
                 .await
                 .unwrap();
         let (handle, shutdown) = state_manager
-            .into_polling_task(Duration::from_millis(100))
+            .into_msg_consumer_task(Duration::from_millis(100))
             .await;
         (handle, shutdown)
     }
