+++ core/src/verifier.rs
@@ -24,8 +24,6 @@ use crate::rpc::clementine::{
     DepositParams, KickoffId, NormalSignatureKind, OperatorKeys, TaggedSignature,
     TransactionRequest, WatchtowerKeys,
 };
-use crate::states;
-use crate::states::syncer::add_new_kickoff_machine;
 use crate::states::StateManager;
 use crate::states::{Duty, Owner};
 use crate::tx_sender::{TxDataForLogging, TxSender};
@@ -184,7 +182,7 @@ impl Verifier {
         let mut state_manager =
             StateManager::new(db.clone(), verifier.clone(), config.protocol_paramset()).await?;
         state_manager.load_from_db().await?;
-        let state_manager_block_syncer = states::syncer::fetch_new_blocks::<Self>(
+        let state_manager_block_syncer = StateManager::<Self>::block_fetcher_task(
             state_manager.get_last_processed_block_height(),
             db.clone(),
             Duration::from_secs(1),
@@ -193,7 +191,7 @@ impl Verifier {
         .await;
 
         let (state_manager_run_loop, shutdown_tx) = state_manager
-            .into_polling_task(Duration::from_secs(1))
+            .into_msg_consumer_task(Duration::from_secs(1))
             .await;
         verifier.state_manager_shutdown_tx = shutdown_tx.into();
 
@@ -361,7 +359,7 @@ impl Verifier {
             reimburse_addr: wallet_reimburse_address,
         };
 
-        states::syncer::add_new_round_machine::<Self>(
+        StateManager::<Self>::dispatch_new_round_machine(
             self.db.clone(),
             &mut dbtx,
             operator_data,
@@ -1141,7 +1139,7 @@ impl Owner for Verifier {
                 if let Some((deposit_data, kickoff_id, _)) = kickoff_data {
                     // add kickoff machine if there is a new kickoff
                     let mut dbtx = self.db.begin_transaction().await?;
-                    add_new_kickoff_machine::<Self>(
+                    StateManager::<Self>::dispatch_new_kickoff_machine(
                         self.db.clone(),
                         &mut dbtx,
                         kickoff_id,
